# Add/sparse waverange optim