<!-- Please be sure to check out our contributing guidelines,
https://github.com/radis/radis/blob/develop/CONTRIBUTING.md . -->

### Description

Implement an major optimization to improve performances when the waverange grid is sparse, i.e., lines or groups of lines are very separated from one another   (related to #222 ) 

Optimization is so far only implemetned in DIT/Voigt mode, but could be easily added in DIT/FFT too   @dcmvdbekerom 


<!-- Provide a general description of what your pull request does. -->

Basically, we mask the waverange to remove regions one truncatino away from any line, before applying the convolution. Note that in the current implementation each (wG, wL) point in the DIT grid can have its own truncation; which is something we want to keep in the future  (if we end up having truncations that are a multiple of (wG, wL) instead of having a common value. 

```

import astropy.units as u
from radis import SpectrumFactory, config, plot_diff

print("Without Sparse Range Optimization")
print("=================================")
print("\n"*2)

config["SPARSE_WAVERANGE"] = False
sf = SpectrumFactory(1, 3000, # 26777,         # cm-1
                  molecule='NO',
                  isotope='1,2,3',
                  pressure=760 * u.torr,
                  verbose=False,
                  wstep=0.01,
                  mole_fraction=1)

sf.fetch_databank('hitemp')
s = sf.eq_spectrum(700)
s.print_perf_profile()


print("With Sparse Range Optimization")
print("==============================")
print("\n"*2)

config["SPARSE_WAVERANGE"] = True
s2 = sf.eq_spectrum(700)
s2.print_perf_profile()

# Plot
s.name = f'0.10.3 ({s.conditions["calculation_time"]:.0f}s)'
s2.name = f'Sparse optim ({s2.conditions["calculation_time"]:.0f}s)'
_, [ax0, ax1] = plot_diff(s, s2)
ax0.set_yscale('log')
```

Output : 

```
Without Sparse Range Optimization
=================================
NOspectrum1533317963200 profiler :
    spectrum_calculation      22.242s ████████████████
        check_line_databank             0.001s 
        reinitialize                    0.007s 
            copy_database                   0.005s 
            memory_usage_warning            0.002s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.026s 
        calc_lineshift                  0.003s 
        calc_hwhm                       0.058s 
        generate_wavenumber_arrays      0.001s 
        calc_line_broadening            22.100s ███████████████
            precompute_DLM_lineshapes       1.070s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.042s 
            DLM_Distribute_lines            0.540s 
            DLM_convolve                    20.306s ██████████████
            others                          0.142s 
        calc_other_spectral_quan        0.040s 
        generate_spectrum_obj           0.006s 
        others                          0.142s 

With Sparse Range Optimization
==============================
NOspectrum1538236138432 profiler :
    spectrum_calculation      5.362s ████████████████
        check_line_databank             0.001s 
        reinitialize                    0.013s 
            copy_database                   0.011s 
            memory_usage_warning            0.002s 
            reset_population                0.000s 
        scaled_eq_linestrength          0.033s 
        calc_lineshift                  0.004s 
        calc_hwhm                       0.068s 
        generate_wavenumber_arrays      0.003s 
        calc_line_broadening            5.205s ███████████████
            precompute_DLM_lineshapes       1.070s ███
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.036s 
            DLM_Distribute_lines            0.486s █
            DLM_convolve                    3.477s ██████████               # ⭐⭐⭐
            others                          0.136s 
        calc_other_spectral_quan        0.030s 
        generate_spectrum_obj           0.004s 
        others                          0.136s
```

So convolution step is 5.7 times faster ; and overall spectrum calculation is 4x faster. The more sparse the better.

![image](https://user-images.githubusercontent.com/16088743/138613345-c6a5174b-8078-4493-ba0d-d12f72515f94.png)

Spectra are virtually the same ; the only difference is only numerical noise (see the diff graph )


Thanks @maydixit  for the help <3