@anandxkumar 2021-10-30
@erwanp, did a benchmark which involved decreasing the `wstep`, and check `spectral points` vs `calculation time` for:
`DLM>FFT`, `DLM>Voigt>Sparse_True`, `DLM>Voigt>Sparse_False`.

**conditions:**
```
# Computation parameters
wmin = 5000
wmax = 5200
T = 300.0 #K
p = 1 #bar
truncation=300

sf = SpectrumFactory(wavenum_min=wmin, wavenum_max=wmax , 
                  pressure=p,
                  truncation=truncation, 
                  molecule="CO",
                  cutoff=0, # 1e-27,
                  verbose=0,
                  )
```
          
Following is the [link](https://anandxkumar.github.io/Benchmark_Visualization_GSoC_2021/LDM%20UPDATED/Wstep(New)/Spectral%20points(Wstep)%20vs%20Calculation%20Time.html).
Atleast for this scenario, I can say Performance of `Sparse_True` vs `Sparse_False` is almost on par with each other. For some cases `Sparse_True` is faster and for others `Sparse_False` is greater. 

So we can set the `config["SPARSE_WAVERANGE"] = True` as default, if it's not harming the accuracy?