@erwanp 2021-10-30
Perfect, thank you for testing. Merging without any worry about performances in current cases :)